+++ src/Processor/Other/Object/PropertyDependentProcessor.php
@@ -10,8 +10,6 @@ use ExtendsSoftware\ExaPHP\Processor\Other\Object\Properties\Property;
 use ExtendsSoftware\ExaPHP\Processor\ProcessorInterface;
 use ExtendsSoftware\ExaPHP\Processor\Result\ResultInterface;
 
-use function array_key_exists;
-
 class PropertyDependentProcessor extends AbstractProcessor
 {
     /**
@@ -31,25 +29,18 @@ class PropertyDependentProcessor extends AbstractProcessor
     /**
      * Properties.
      *
-     * @var array<int|string, Property>
+     * @var array<mixed, Property>
      */
     private array $properties = [];
 
     /**
      * PropertyDependentProcessor constructor.
      *
-     * @param string                                     $property
-     * @param array<int|string, ProcessorInterface>|null $processors
-     * @param bool|null                                  $strict
+     * @param string    $property
+     * @param bool|null $strict
      */
-    public function __construct(
-        private readonly string $property,
-        ?array $processors = null,
-        private readonly ?bool $strict = true
-    ) {
-        foreach ($processors ?? [] as $name => $processor) {
-            $this->addProperty($name, $processor);
-        }
+    public function __construct(private readonly string $property, private readonly ?bool $strict = true)
+    {
     }
 
     /**
@@ -69,14 +60,15 @@ class PropertyDependentProcessor extends AbstractProcessor
             return $this->getValidResult($value);
         }
 
-        $dependent = $context->{$this->property};
-        if (!isset($this->properties[$dependent])) {
-            if (array_key_exists('*', $this->properties)) {
-                $dependent = '*';
-            } else {
+        $propertyValue = $context->{$this->property};
+        $property = $this->findProperty($propertyValue);
+
+        if (!$property) {
+            $property = $this->findProperty('*');
+            if (!$property) {
                 if ($this->strict) {
                     return $this->getInvalidResult(self::PROCESSOR_MISSING, [
-                        'property' => $dependent,
+                        'property' => $propertyValue,
                     ]);
                 }
 
@@ -84,7 +76,7 @@ class PropertyDependentProcessor extends AbstractProcessor
             }
         }
 
-        return $this->properties[$dependent]->getProcessor()->process($value, $context);
+        return $property->getProcessor()->process($value, $context);
     }
 
     /**
@@ -92,16 +84,34 @@ class PropertyDependentProcessor extends AbstractProcessor
      *
      * An existing processor for property value will be overwritten.
      *
-     * @param int|string         $name
+     * @param mixed              $value
      * @param ProcessorInterface $processor
      *
      * @return PropertyDependentProcessor
      */
-    public function addProperty(int|string $name, ProcessorInterface $processor): PropertyDependentProcessor
+    public function addProperty(mixed $value, ProcessorInterface $processor): PropertyDependentProcessor
     {
-        $this->properties[$name] = new Property($name, $processor);
+        $this->properties[$value] = new Property($value, $processor);
 
         return $this;
+    }
+
+    /**
+     * Find property for value.
+     *
+     * @param mixed $value
+     *
+     * @return Property|null
+     */
+    protected function findProperty(mixed $value): ?Property
+    {
+        foreach ($this->properties as $property) {
+            if ($property->getValue() === $value) {
+                return $property;
+            }
+        }
+
+        return null;
     }
 
     /**