+++ tests/Processor/Other/Object/PropertiesProcessorTest.php
@@ -42,12 +42,11 @@ class PropertiesProcessorTest extends TestCase
             ->method('process')
             ->willReturn($innerResult);
 
-        $processor = new PropertiesProcessor([
-            'foo' => $innerProcessor,
-            'bar' => $innerProcessor,
-            'baz' => $innerProcessor,
-            'qux' => new ProxyProcessor($innerProcessor),
-        ]);
+        $processor = new PropertiesProcessor();
+        $processor->addProperty('foo', $innerProcessor)
+            ->addProperty('bar', $innerProcessor)
+            ->addProperty('baz', $innerProcessor)
+            ->addProperty('qux', new ProxyProcessor($innerProcessor));
         $result = $processor->process($object, 'context');
 
         $this->assertInstanceOf(ObjectContainerResult::class, $result);
@@ -84,9 +83,8 @@ class PropertiesProcessorTest extends TestCase
             ->with('bar', $object)
             ->willReturn($innerResult);
 
-        $processor = new PropertiesProcessor([
-            'foo' => $innerProcessor,
-        ]);
+        $processor = new PropertiesProcessor();
+        $processor->addProperty('foo', $innerProcessor);
         $result = $processor->process($object, 'context');
 
         $this->assertInstanceOf(ObjectContainerResult::class, $result);
@@ -132,9 +130,9 @@ class PropertiesProcessorTest extends TestCase
             ->expects($this->never())
             ->method('process');
 
-        $processor = new PropertiesProcessor([
-            'foo' => $innerProcessor,
-        ]);
+        $processor = new PropertiesProcessor();
+        $processor->addProperty('foo', $innerProcessor);
+
         $result = $processor->process((object)[], 'context');
 
         $this->assertInstanceOf(ObjectContainerResult::class, $result);
@@ -183,9 +181,8 @@ class PropertiesProcessorTest extends TestCase
             ->method('process')
             ->willReturn($innerResult);
 
-        $processor = new PropertiesProcessor([
-            'foo' => $innerProcessor,
-        ]);
+        $processor = new PropertiesProcessor();
+        $processor->addProperty('foo', $innerProcessor);
 
         $result1 = $processor->process($object);
         $result2 = $processor->process($object, 'foo');