+++ tests/Processor/Other/Object/PropertyDependentProcessorTest.php
@@ -13,15 +13,32 @@ use PHPUnit\Framework\TestCase;
 class PropertyDependentProcessorTest extends TestCase
 {
     /**
+     * Data provider for is valid.
+     *
+     * @return array[]
+     */
+    public static function isValidProvider(): array
+    {
+        return [
+            ['foo'],
+            [true],
+            [null],
+            [2],
+        ];
+    }
+
+    /**
      * Valid.
      *
      * Test that an object is valid.
      *
-     * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::__construct()
-     * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::addProperty()
-     * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::process()
+     * @dataProvider isValidProvider()
+     * @covers       \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::__construct()
+     * @covers       \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::addProperty()
+     * @covers       \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::process()
+     * @covers       \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::findProperty()
      */
-    public function testIsValid(): void
+    public function testIsValid(mixed $value): void
     {
         $innerResult = $this->createMock(ResultInterface::class);
         $innerResult
@@ -40,14 +57,13 @@ class PropertyDependentProcessorTest extends TestCase
             ->expects($this->never())
             ->method('process');
 
-        $processor = new PropertyDependentProcessor('foo', [
-            'bar' => $innerProcessor1,
-            'baz' => $innerProcessor2,
-        ]);
+        $processor = (new PropertyDependentProcessor('foo'))
+            ->addProperty($value, $innerProcessor1)
+            ->addProperty('baz', $innerProcessor2);
         $result = $processor->process(
             'qux',
             (object)[
-                'foo' => 'bar',
+                'foo' => $value,
             ],
         );
 
@@ -62,6 +78,7 @@ class PropertyDependentProcessorTest extends TestCase
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::__construct()
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::addProperty()
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::process()
+     * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::findProperty()
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::getTemplates()
      */
     public function testStrictAndContextPropertyMissing(): void
@@ -76,10 +93,9 @@ class PropertyDependentProcessorTest extends TestCase
             ->expects($this->never())
             ->method('process');
 
-        $processor = new PropertyDependentProcessor('qux', [
-            'bar' => $processor1,
-            'baz' => $processor2,
-        ]);
+        $processor = (new PropertyDependentProcessor('qux'))
+            ->addProperty('bar', $processor1)
+            ->addProperty('baz', $processor2);
         $result = $processor->process(
             'qux',
             (object)[
@@ -98,6 +114,7 @@ class PropertyDependentProcessorTest extends TestCase
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::__construct()
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::addProperty()
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::process()
+     * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::findProperty()
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::getTemplates()
      */
     public function testStrictAndProcessorMissing(): void
@@ -112,10 +129,9 @@ class PropertyDependentProcessorTest extends TestCase
             ->expects($this->never())
             ->method('process');
 
-        $processor = new PropertyDependentProcessor('foo', [
-            'qux' => $innerProcessor1,
-            'baz' => $innerProcessor2,
-        ]);
+        $processor = (new PropertyDependentProcessor('foo'))
+            ->addProperty('qux', $innerProcessor1)
+            ->addProperty('baz', $innerProcessor2);
         $result = $processor->process(
             'qux',
             (object)[
@@ -147,10 +163,9 @@ class PropertyDependentProcessorTest extends TestCase
             ->expects($this->never())
             ->method('process');
 
-        $processor = new PropertyDependentProcessor('qux', [
-            'bar' => $innerProcessor1,
-            'baz' => $innerProcessor2,
-        ], false);
+        $processor = (new PropertyDependentProcessor('qux', false))
+            ->addProperty('bar', $innerProcessor1)
+            ->addProperty('baz', $innerProcessor2);
         $result = $processor->process(
             'qux',
             (object)[
@@ -170,6 +185,7 @@ class PropertyDependentProcessorTest extends TestCase
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::__construct()
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::addProperty()
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::process()
+     * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::findProperty()
      */
     public function testNonStrictAndProcessorMissing(): void
     {
@@ -183,10 +199,9 @@ class PropertyDependentProcessorTest extends TestCase
             ->expects($this->never())
             ->method('process');
 
-        $processor = new PropertyDependentProcessor('foo', [
-            'qux' => $innerProcessor1,
-            'baz' => $innerProcessor2,
-        ], false);
+        $processor = (new PropertyDependentProcessor('foo', false))
+            ->addProperty('qux', $innerProcessor1)
+            ->addProperty('baz', $innerProcessor2);
         $result = $processor->process(
             'qux',
             (object)[
@@ -206,6 +221,7 @@ class PropertyDependentProcessorTest extends TestCase
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::__construct()
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::addProperty()
      * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::process()
+     * @covers \ExtendsSoftware\ExaPHP\Processor\Other\Object\PropertyDependentProcessor::findProperty()
      */
     public function testMissingPropertyAndWildcardProcessor(): void
     {
@@ -221,9 +237,8 @@ class PropertyDependentProcessorTest extends TestCase
             ->with('qux')
             ->willReturn($innerResult);
 
-        $processor = new PropertyDependentProcessor('foo', [
-            '*' => $innerProcessor,
-        ], false);
+        $processor = (new PropertyDependentProcessor('foo', false))
+            ->addProperty('*', $innerProcessor);
         $result = $processor->process(
             'qux',
             (object)[